+++ changelog.md
@@ -2,6 +2,8 @@ Codion Change Log
 ==================
 
 ## 0.17.36-SNAPSHOT
+### is.codion.framework.model
+- EntityEditModel.delete() now returns the deleted entities.
 
 ## 0.17.35
 ### is.codion.common.model